refactor: http 구성 요소 패키지 이동
