+++ tomcat/src/main/java/org/apache/coyote/component/HttpCookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.component;
 
 import java.util.Collections;
 import java.util.HashMap;
