+++ tomcat/src/main/java/org/apache/coyote/component/HttpHeaderField.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.component;
 
 public enum HttpHeaderField {
