+++ tomcat/src/main/java/org/apache/coyote/component/HttpHeaders.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.component;
 
 import java.io.BufferedReader;
 import java.io.IOException;
