+++ tomcat/src/main/java/org/apache/coyote/component/HttpMethod.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.component;
 
 import java.util.Arrays;
 import java.util.Map;
