+++ tomcat/src/main/java/org/apache/coyote/component/HttpStatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.component;
 
 public enum HttpStatusCode {
