+++ tomcat/src/main/java/org/apache/coyote/controller/DefaultController.java
@@ -4,7 +4,7 @@ import org.apache.coyote.Controller;
 import org.apache.coyote.http11.AbstractController;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
-import org.apache.coyote.http11.HttpStatusCode;
+import org.apache.coyote.component.HttpStatusCode;
 
 public class DefaultController extends AbstractController implements Controller {
