+++ tomcat/src/test/java/org/apache/coyote/http11/ContentTypeTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.Stream;
 
+import org.apache.coyote.component.ContentType;
+import org.apache.coyote.component.Path;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
